requirements.txt
streamlit==1.31.0
google-generativeai
python-dotenv==1.0.1
Pillow
PyPDF2
streamlit-paste-button
